Changed a bit

+++ helpers/environment.go
@@ -2,6 +2,6 @@ package helpers
 
 import "fmt"
 
-func GetPort(portValue int64) string {
+func GetPort(portValue uint16) string {
 	return fmt.Sprintf(":%d", portValue)
 }
